Move out array view methods
